GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - Enhancements (refs/gitbutler/Enhancements)
   - src/components/Loading.jsx
   - src/styles/Navbar.css
   - src/components/Navbar.jsx
   - src/App.jsx
   - src/pages/LogIn.jsx
   - src/components/PrivateRoute.jsx
   - src/pages/Post.jsx
   - src/context/AuthProvider.jsx
   - src/components/PageScaffold.jsx
   - src/pages/Profile.jsx
   - src/pages/Resume.jsx
   - src/pages/Posts.jsx
   - src/components/Redirect.jsx
   - src/pages/PostForm.jsx
   - src/pages/Contact.jsx
   - src/components/PostListing.jsx
   - src/pages/Spinner.jsx
   - src/components/Spinner.jsx

Your previous branch was: refs/heads/master

The sha for that commit was: 30fa1ef0cb9813573cabd7203bb6d1853ffade80

For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch